requirements.txt
streamlit
pandas
google-cloud-bigquery
google-generativeai
db-dtypes
plotly
geopy
prophet